server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
sse-starlette==1.8.2

# Celery monitoring dependencies
kombu==5.3.4
psutil==5.9.8
pydantic-settings==2.1.0
flower==2.0.1

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0

# Development dependencies
pre-commit==3.6.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
bandit==1.7.5

# Monitoring dependencies
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0